API

Components listing

Modules

MGVI.MGVI

Types and constants

MGVI.AbstractJacobianFunc

MGVI.AbstractResidualSampler

MGVI.FullJacobianFunc

MGVI.FullResidualSampler

MGVI.FwdDerJacobianFunc

MGVI.FwdRevADJacobianFunc

MGVI.ImplicitResidualSampler

Functions and macros

MGVI.mgvi_kl_optimize_step

Documentation

MGVI.MGVI — Module

MGVI

An implementation of the Metric Gaussian Variational Inference algorithm.

MGVI.FullJacobianFunc — Type

FullJacobianFunc(f)

Construct of the Jacobian with ForwardDiff.

When called at point θ, the Jacobian matrix is fully instantiated and stored explicitly in memory.

Examples

# forward_model: θ -> Distribution
jacobian_func = FullJacobianFunc(forward_model)
jacobian = jacobian_func(θ)  # LinearMap
jacobian * v  # act on vector

MGVI.FullResidualSampler — Type

FullResidualSampler(λ_information, jac_dλ_dθ)

Gaussian posterior's covariance approximated by the Fisher Information

Caution: This sampler constructs the covariance matrix explicitly, so memory use grows quadratically with the number of model parameters

The Fisher information in canonical coordinates and Jacobian of the coordinate transformation are provided as arguments. Since $J \cdot I_{canon} \cdot J^T$ does not depend on the definition of canonical coordinates, we omit description of what canonical coordinates are.

Arguments

λ_information::LinearMap: Fisher Information in canonical coordinates. Coordinates are chosen in the way that this matrix is very simple, and diagonal for the univariate distributions

jac_dλ_fθ::LinearMap: Coordinate transformation jacobian between canonical coordinates and coordinates of the model

MGVI.FwdDerJacobianFunc — Type

FwdDerJacobianFunc(f)

Construct of the Jacobian with ForwardDiff for the direct action, and twice applied ForwardDiff for the adjoint action

Adjoint is implemented by introducing a placeholder parametric vector $\vec{t}$:

$\frac{d}{d\vec{t}} \vec{x} \cdot (A \vec{t}) = A^{T} \vec{x}$

This allows to implement both, the direct and adjoint actions, using ForwardDiff, without instantiating full Jacobian at any point.

MGVI.FwdRevADJacobianFunc — Type

FwdRevADJacobianFunc(f)

Construct of the Jacobian with ForwardDiff for direct action and Zygote.pullback for the adjoint

The Jacobian action is computed on the fly, no matrix is stored in memory at any time.

Examples

# forward_model: θ -> Distribution
jacobian_func = FwdRevADJacobianFunc(forward_model)
jacobian = jacobian_func(θ)  # LinearMap
jacobian * v  # act on vector as if it was a matrix

MGVI.ImplicitResidualSampler — Type

ImplicitResidualSampler(λ_information, jac_dλ_dθ; cg_params=NamedTuple())

Memory efficient Gaussian posterior's covariance approximated by the Fisher Information

This sampler uses Conjugate Gradients to iteratively invert the Fisher information, never instantiating the covariance in memory explicitly.

Arguments

λ_information::LinearMap: Fisher Information in canonical coordinates.

jac_dλ_fθ::LinearMap: Coordinate transformation jacobian between canonical coordinates and

cg_params::NamedTuple: Keyword arguments passed to cg. Useful for enabling verbose mode or to set accuracy/number of iterations

MGVI.mgvi_kl_optimize_step — Method

mgvi_kl_optimize_step(rng, forward_model, data, init_param_point;
                      jacobian_func=jacobian_func,
                      residual_sampler=residual_sampler,
                      [num_residuals=3,]
                      [residual_sampler_options=NamedTuple(),]
                      [optim_solver=LBFGS(),]
                      [optim_options=Optim.Options()])

Performs one MGVI iteration.

The posterior distribution is approximated with a multivariate normal distribution. The covariance is approximated with the inverse Fisher information valuated at init_param_point. Samples are drawn according to this covariance, which are then used to estimate and minimize the KL divergence between the true posterior and the approximation.

Arguments

rng::AbstractRNG: instance of the random number generator

forward_model::Function: turns model parameters into an instance of Distribution

data::AbstractVector: data on which model's pdf is evaluated

init_param_point::Vector: initial estimate of the model parameters

jacobian_func::Type{<:AbstractJacobianFunc}: method to calculate the Jacobian of the forward model

residual_sampler::Type{<:AbstractResidualSampler}: method to draw samples from the approximation

num_residuals::Integer = 3: number of samples used to estimate the KL divergence

residual_sampler_options::NamedTuple = NamedTuple(): further options to pass to the residual sampler. Important is cg_params that is passed to the CG solver used inside of ImplicitResidualSampler

optim_solver::Optim.AbstractOptimizer = LBFGS(): optimizer used for minimizing KL divergence

optim_options::Optim.Options = Optim.Options(): options to pass to the KL optimizer

Example

using Random, Distributions, MGVI

model(x::AbstractVector) = Normal(x[1], 0.2)
true_param = [2.0]
data = rand(model(true_param), 1)[1]
init_param = [1.3]

res = mgvi_kl_optimize_step(Random.GlobalRNG, model, data, init_param;
                            jacobian_func=FwdRevADJacobianFunc,
                            residual_sampler=ImplicitResidualSampler,
                            num_residuals=5,
                            residual_sampler_options=(;cg_params=(;maxiter=10, verbose=true))),
                            optim_solver=LBFGS(;m=5),
                            optim_options=Optim.Options(iterations=7, show_trace=true))

next_param_point = res.result

optim_optimized_object = res.optimized
Optim.summary(optim_optimized_object)

samples_from_est_covariance = res.samples

See also

Residual samplers: AbstractResidualSampler, ImplicitResidualSampler, FullResidualSampler

Jacobian functions: AbstractJacobianFunc, FwdRevADJacobianFunc, FwdDerJacobianFunc

MGVI.AbstractJacobianFunc — Type

abstract type AbstractJacobianFunc <: Function end

Abstract type of the jacobian calculators.

Represents the jacobian of the function. The instance can be called at the point θ, a LinearMap representing the jacobian is returned.

MGVI.AbstractResidualSampler — Type

abstract type AbstractResidualSampler <: Sampleable{Multivariate, Continuous} end

Generate zero-mean samples from Gaussian posterior, assuming priors are standard gaussians.

Example

rng = MersenneTwister(145)
rs = FullResidualSampler(λ_information, jac_dλ_dθ)
samples = rand(rng, rs, 10)  # generate 10 samples