API

Components listing

Modules

MGVI.MGVI

Types and constants

MGVI.LinearSolverAlg

MGVI.MGVIContext

MGVI.MatrixInversion

MGVI.NewtonCG

MGVI.PDLinMapWithChol

MGVI.ResidualSampler

Functions and macros

MGVI.fisher_information

MGVI.mgvi_optimize_step

Documentation

MGVI.MGVI — Module

MGVI

An implementation of the Metric Gaussian Variational Inference algorithm.

MGVI.MGVIContext — Type

MGVIContext{OP=LinearMap}(rng::AbstractRNG, ad::AutoDiffOperators.ADSelector)

Specifies the linear operator type, RNG and automatic differentiation backend to be used by MGVI operations.

MGVI.ResidualSampler — Type

struct ResidualSampler

Generates zero-centered samples from the posterior's covariance approximated by the Fisher information.

This sampler uses Conjugate Gradients to iteratively invert the Fisher information, never instantiating the covariance in memory explicitly.

The Fisher information in canonical coordinates and Jacobian of the coordinate transformation are provided as arguments.

Constructor:

ResidualSampler(f_model::Function, center_point::Vector{<:Real}, solver::MGVI.LinearSolverAlg, context::MGVIContext)

Call MGVI.sample_residuals(s::ResidualSampler[, n::Integer]) to generate a single or n samples.

MGVI.mgvi_optimize_step — Function

function mgvi_optimize_step(
    forward_model::Function, data, init_param_point::AbstractVector{<:Real}, context::MGVIContext;
    num_residuals::Integer = 3,
    lcenter_pointinear_solver::LinearSolverAlg = IterativeSolversCG(),
    optim_solver::Union{Optim.AbstractOptimizer,MGVI.NewtonCG} = MGVI.NewtonCG(),
    optim_options::Union{Optim.Options, Nothing} = Optim.Options(),
)

Performs one MGVI iteration.

The posterior distribution is approximated with a multivariate normal distribution. The covariance is approximated with the inverse Fisher information valuated at init_param_point. Samples are drawn according to this covariance, which are then used to estimate and minimize the KL divergence between the true posterior and the approximation.

Note: The prior is implicit, it is a standard (uncorrelated) multivariate normal distribution of the same dimensionality as init_param_point.

Example

using Random, Distributions, MGVI
import Zygote

context = MGVIContext(ADModule(:Zygote))

model(x::AbstractVector) = Normal(x[1], 0.2)
true_param = [2.0]
data = rand(model(true_param), 1)[1]
init_param = [1.3]

res = mgvi_optimize_step(
    model, data, init_param, context;
    num_residuals = 5,
    linear_solver = MGVI.IterativeSolversCG(),
    optim_solver = MGVI.NewtonCG(),
)

next_param_point = res.result

optim_optimized_object = res.optimized
Optim.summary(optim_optimized_object)

samples_from_est_covariance = res.samples

MGVI.LinearSolverAlg — Type

abstract type LinearSolverAlg

Abstract supertype for linear solver algorithms.

MGVI.MatrixInversion — Type

struct MatrixInversion <: LinearSolverAlg

Solve linear systems by direct matrix inversion.

Note: Will instantiate implicit matrices/operators in memory explicitly.

MGVI.NewtonCG — Type

struct NewtonCG

Constructors:

'''NewtonCG(; fields...)'''

α::Float64

amount of previous NewtonCG improvement guarding the lower bound to the improvement between consecutive cg iterations from the second NewtonCG step on Default: 0.1

steps::Int64

Number of total NewtonCG steps Default: 4

i₀::Int64

maximum number of cg iterations in the first NewtonCG step Default: 5

i₁::Int64

maximum number of cg iterations from the second NewtonCG step on Default: 50

linesearcher::Any

LineSearcher that will be used after cg iterations are finished Default: StrongWolfe{Float64}()

MGVI.PDLinMapWithChol — Type

struct MGVI.PDLinMapWithChol{T} <: LinearMaps.LinearMap{T}

A LinearMap that stores both a map and the lower-tringangular map of its Cholesky decomposition.

MGVI.fisher_information — Function

MGVI.fisher_information(distribution::Distributions.Distribution)

Get the fisher information matrix/operator (as a LinearMap) for the given distribution.